[refactor](SqlManagement): Table action layout and remark column[skip ci]
