+++ packages/sqle/src/page/SqlManagement/component/SQLEEIndex/column.tsx
@@ -15,13 +15,11 @@ import { tooltipsCommonProps } from '@actiontech/shared/lib/components/BasicTool
 import { Avatar } from 'antd';
 import StatusTag from './StatusTag';
 import { BasicTag, BasicTypographyEllipsis } from '@actiontech/shared';
-// import { ACTIONTECH_TABLE_ACTION_BUTTON_WIDTH } from '@actiontech/shared/lib/components/ActiontechTable/hooks/useTableAction';
 import { SQLAuditRecordListUrlParamsKey } from './index.data';
 import { SqlManageAuditStatusEnum } from '@actiontech/shared/lib/api/sqle/service/common.enum';
-import { SupportLanguage } from '@actiontech/shared/lib/enum';
 import {
-  ActiontechTableActionsWithPermissions,
-  PERMISSIONS
+  PERMISSIONS,
+  PermissionsConstantType
 } from '@actiontech/shared/lib/global';
 
 export type SqlManagementTableFilterParamType = PageInfoWithoutIndexAndSize<
@@ -106,96 +104,18 @@ export const ExtraFilterMeta: () => ActiontechTableFilterMeta<
   ]);
 };
 
-export const SqlManagementRowAction = (
-  openModal: (name: ModalName, row?: ISqlManage) => void,
-  jumpToAnalyze: (sqlManageID: string) => void,
-  openCreateSqlManagementExceptionModal: (record?: ISqlManage) => void,
-  onCreateWhitelist: (record?: ISqlManage) => void,
-  language: SupportLanguage
-): ActiontechTableActionsWithPermissions<ISqlManage> => {
-  // const getWidth = () => {
-  //   if (operationPermission) {
-  //     return language === SupportLanguage.enUS
-  //       ? 350
-  //       : ACTIONTECH_TABLE_ACTION_BUTTON_WIDTH * 3;
-  //   }
-  //   return 110;
-  // };
-
-  return {
-    // todo width根据权限和语言适配
-    // width: getWidth(),
-    buttons: [
-      {
-        text: t('sqlManagement.table.action.single.assignment'),
-        key: 'assignment-single',
-        buttonProps: (record) => {
-          return {
-            onClick: () => {
-              openModal(ModalName.Assignment_Member_Single, record);
-            }
-          };
-        },
-        permissions: PERMISSIONS.ACTIONS.SQLE.SQL_MANAGEMENT.ASSIGNMENT
-      },
-      {
-        text: t('sqlManagement.table.action.single.updateStatus.triggerText'),
-        key: 'change-status-single',
-        buttonProps: (record) => {
-          return {
-            onClick: () => {
-              openModal(ModalName.Change_Status_Single, record);
-            }
-          };
-        },
-        permissions: PERMISSIONS.ACTIONS.SQLE.SQL_MANAGEMENT.UPDATE_STATUS
-      }
-    ],
-    moreButtons: [
-      {
-        text: t('sqlManagement.table.action.single.updatePriority.triggerText'),
-        key: 'change-priority-single',
-        onClick: (record) => {
-          openModal(ModalName.Change_SQL_Priority, record);
-        },
-        permissions: PERMISSIONS.ACTIONS.SQLE.SQL_MANAGEMENT.UPDATE_PRIORITY
-      },
-      {
-        text: t('sqlManagement.table.action.analyze'),
-        key: 'analyze-sql',
-        onClick: (record) => {
-          jumpToAnalyze(record?.id?.toString() ?? '');
-        }
-      },
-      {
-        text: t('sqlManagement.table.action.createSqlManagementException'),
-        key: 'create-exception',
-        onClick: (record) => {
-          openCreateSqlManagementExceptionModal(record);
-        },
-        permissions:
-          PERMISSIONS.ACTIONS.SQLE.SQL_MANAGEMENT.CREATE_SQL_EXCEPTION
-      },
-      {
-        text: t('sqlManagement.table.action.createWhitelist'),
-        key: 'create-whitelist',
-        onClick: (record) => {
-          onCreateWhitelist(record);
-        },
-        permissions: PERMISSIONS.ACTIONS.SQLE.SQL_MANAGEMENT.CREATE_WHITE_LIST
-      }
-    ]
-  };
-};
-
 const SqlManagementColumn: (
   projectID: string,
   updateRemark: (id: number, remark: string) => void,
-  openModal: (name: ModalName, row?: ISqlManage) => void
+  openModal: (name: ModalName, row?: ISqlManage) => void,
+  checkActionPermission: (
+    requiredPermission: PermissionsConstantType
+  ) => boolean
 ) => ActiontechTableColumn<ISqlManage, SqlManagementTableFilterParamType> = (
   projectID,
   updateRemark,
-  openModal
+  openModal,
+  checkActionPermission
 ) => {
   return [
     {
@@ -415,10 +335,11 @@ const SqlManagementColumn: (
       width: 200,
       className: 'ellipsis-column-width',
       render: (remark, record) => {
-        // todo column中的权限判断 admin project_manager project_not_archived
-        // todo 是否可以从usePermission中导出一个通用的权限 这个权限感觉是最常用的 （（ admin || project_manager） && project_not_archived）
-        const hasPermissionAndNotArchive = true;
-        if (!hasPermissionAndNotArchive)
+        if (
+          !checkActionPermission(
+            PERMISSIONS.ACTIONS.SQLE.SQL_MANAGEMENT.EDIT_REMARK
+          )
+        )
           return remark ? <BasicTypographyEllipsis textCont={remark} /> : '-';
         return (
           <EditText
