+++ packages/sqle/src/page/SqlManagement/component/SQLEEIndex/index.tsx
@@ -33,9 +33,13 @@ import {
 } from '@actiontech/shared/lib/api/sqle/service/SqlManage/index.enum';
 import SqlManagementColumn, {
   ExtraFilterMeta,
-  SqlManagementRowAction,
   type SqlManagementTableFilterParamType
 } from './column';
+import {
+  SqlManagementRowAction,
+  actionsButtonData,
+  defaultActionButton
+} from './actions';
 import { ModalName } from '../../../../data/ModalName';
 import { SorterResult, TableRowSelection } from 'antd/es/table/interface';
 import { ISqlManage } from '@actiontech/shared/lib/api/sqle/service/common';
@@ -46,7 +50,6 @@ import EventEmitter from '../../../../utils/EventEmitter';
 import { DB_TYPE_RULE_NAME_SEPARATOR } from './hooks/useRuleTips';
 import useSqlManagementRedux from './hooks/useSqlManagementRedux';
 import useBatchIgnoreOrSolve from './hooks/useBatchIgnoreOrSolve';
-import { actionsButtonData, defaultActionButton } from './index.data';
 import useGetTableFilterInfo from './hooks/useGetTableFilterInfo';
 import { DownArrowLineOutlined } from '@actiontech/icons';
 import useSqlManagementExceptionRedux from '../../../SqlManagementException/hooks/useSqlManagementExceptionRedux';
@@ -57,8 +60,11 @@ import { SqlManageAuditStatusEnum } from '@actiontech/shared/lib/api/sqle/servic
 const SQLEEIndex = () => {
   const { t } = useTranslation();
   const [messageApi, messageContextHolder] = message.useMessage();
-  const { parse2TableActionPermissions, parse2TableToolbarActionPermissions } =
-    usePermission();
+  const {
+    parse2TableActionPermissions,
+    parse2TableToolbarActionPermissions,
+    checkActionPermission
+  } = usePermission();
   // api
   const { projectID, projectName } = useCurrentProject();
   const { username, userId, language } = useCurrentUser();
@@ -232,7 +238,8 @@ const SQLEEIndex = () => {
         jumpToAnalyze,
         onCreateSqlManagementException,
         onCreateWhitelist,
-        language
+        language,
+        checkActionPermission
       )
     );
   }, [
@@ -241,7 +248,8 @@ const SQLEEIndex = () => {
     onCreateSqlManagementException,
     onCreateWhitelist,
     language,
-    parse2TableActionPermissions
+    parse2TableActionPermissions,
+    checkActionPermission
   ]);
 
   const [selectedRowKeys, setSelectedRowKeys] = useState<number[]>([]);
@@ -272,8 +280,14 @@ const SQLEEIndex = () => {
   );
 
   const columns = useMemo(
-    () => SqlManagementColumn(projectID, updateRemark, openModal),
-    [projectID, updateRemark, openModal]
+    () =>
+      SqlManagementColumn(
+        projectID,
+        updateRemark,
+        openModal,
+        checkActionPermission
+      ),
+    [projectID, updateRemark, openModal, checkActionPermission]
   );
 
   const tableSetting = useMemo<ColumnsSettingProps>(
